% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calc_stats}
\alias{calc_stats}
\title{Calculate statistics for a given raster.}
\usage{
calc_stats(layercode, raster)
}
\arguments{
\item{layercode}{character.}

\item{raster}{RasterLayer.}
}
\value{
A dataframe with the layercode and all basic statistics.
}
\description{
Internal method used to calculate the statistics of the layers.
}
\seealso{
\code{\link{layer_stats}}
}

