% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{compress_file}
\alias{compress_file}
\title{Compress a file}
\usage{
compress_file(filename, outputdir, method = "gzip", overwrite = FALSE,
remove = TRUE)
}
\arguments{
\item{filename}{character. Path to the file that should be compressed.}

\item{outputdir}{character. Path to the output directory.}

\item{method}{character. The default value is \code{"gzip"}, the supported
methods are \code{"gzip"} and \code{"bzip2"}.}

\item{overwrite}{logical. If \code{TRUE} and the output file already exists
then the file is silently overwritten, otherwise an exception is thrown.}

\item{remove}{logical. If \code{TRUE} then the input file is remove
afterwards, otherwise not.}
}
\value{
The path of the compressed file.
}
\description{
The internal method \code{compress_file} compresses a file and returns the
path to the result.
}
\seealso{
\code{\link{decompress_file}}
}

