% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{equalarea_project}
\alias{equalarea_project}
\title{Project and saves a layer from WGS84 (EPGS:4326) to Behrmann Equal Area (ESRI:54017) coordinate reference system}
\usage{
equalarea_project(path)
}
\arguments{
\item{path}{the path to the WGS84 layer}
}
\value{
the path to the Behrmann Equal Area layer
}
\description{
Project and saves a layer from WGS84 (EPGS:4326) to Behrmann Equal Area (ESRI:54017) coordinate reference system
}
\seealso{
\code{\link{load_layers}}
}
\keyword{internal}
