% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.R
\docType{data}
\name{equalareaproj}
\alias{equalareaproj}
\title{World Behrmann equal area coordinate reference system (ESRI:54017),
used when using load_layers with equal_area = TRUE}
\format{\preformatted{Formal class 'CRS' [package "sp"] with 1 slot
  ..@ projargs: chr "+proj=cea +lon_0=0 +lat_ts=30 +x_0=0 +y_0=0 +datum=WGS84 +ellps=WGS84 +units=m +no_defs +towgs84=0,0,0"
}}
\usage{
equalareaproj
}
\description{
World Behrmann equal area coordinate reference system (ESRI:54017),
used when using load_layers with equal_area = TRUE
}
\keyword{datasets}

