% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get_datadir}
\alias{get_datadir}
\title{Get data directory}
\usage{
get_datadir(datadir)
}
\arguments{
\item{datadir}{character. Directory as will be passed to 
\code{\link{load_layers}}. This does not change the data directory used by 
\code{\link{load_layers}} but only shows the result of passing a specific 
datadir. If \code{NULL} is used then this returns 
\code{file.path(path.expand("~"), "R", "sdmpredictors")}. This can be 
overridden by setting the \code{sdmpredictors_datadir} option with 
\code{options(sdmpredictors_datadir = "<your preferred directory>")}.}
}
\description{
Find out where the environmental data is stored
}
\examples{
get_datadir(NULL)
}
\seealso{
\code{\link{load_layers}}
}
\keyword{internal}

