% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{layer_stats}
\alias{layer_stats}
\title{Gives basic layer statistics}
\usage{
layer_stats(layercodes = c())
}
\arguments{
\item{layercodes}{character vector or dataframe. Codes of the layers you want
the basic statistics of as a character vector or a dataframe with a "layer_code" column.
With the default empty vector all statistics are returned.}
}
\value{
A dataframe with basic statistics about each given layercode.
}
\description{
\code{layer_stats} returns basic statistics (minimum, q1, median, q3, maximum,
median absolute deviation (mad), mean, standard deviation (sd)) about each
given layercode.
}
\examples{
# layer stats for the first 10 layers
layer_stats()[1:10,]
layer_stats(c("BO_calcite", "MS_bathy_5m"))
}
\seealso{
\code{\link{list_layers}} \code{\link{layers_correlation}}
}

