% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list.R
\name{list_layers}
\alias{list_layers}
\title{List the layers provided by one or more datasets}
\usage{
list_layers(datasets=c(), terrestrial = TRUE, marine = TRUE)
}
\arguments{
\item{datasets}{character vector. Code of the datasets.
#'}

\item{terrestrial}{logical. When \code{TRUE}, then datasets
that only have terrestrial data (seamasked) are returned.}

\item{marine}{logical. When \code{TRUE}, then datasets
that only have marine data (landmasked) are returned.

By default it returns all layers from all datasets, when both marine and terrestrial are
\code{FALSE} then only datasets without land- nor seamasks are returned.}
}
\value{
A dataframe with information on the supported datasets.
}
\description{
\code{list_layers} returns information on the
layers of one or more datasets.
}
\examples{
list_layers()
list_layers("WorldClim")
list_layers("Bio-ORACLE")
list_layers(c("Bio-ORACLE","MARSPEC"))
list_layers(marine=FALSE)
list_layers(terrestrial=FALSE)
}
\seealso{
\code{\link{list_datasets}}, \code{\link{load_layers}}
}

