% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.R
\docType{data}
\name{lonlatproj}
\alias{lonlatproj}
\title{Longitude/latitude coordinate reference system (EPSG:4326),
used when using load_layers with equal_area = FALSE}
\format{\preformatted{Formal class 'CRS' [package "sp"] with 1 slot
  ..@ projargs: chr "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
}}
\usage{
lonlatproj
}
\description{
Longitude/latitude coordinate reference system (EPSG:4326),
used when using load_layers with equal_area = FALSE
}
\keyword{datasets}

