% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pearson_correlation}
\alias{pearson_correlation}
\title{Calculate Pearson correlation coefficient for part of a matrix}
\usage{
pearson_correlation(mat, i, j, iR, jR, sds, n, asSample)
}
\description{
Calculate Pearson correlation coefficient for part of a matrix
}
\seealso{
\code{\link{layers_correlation} \link{pearson_correlation_matrix}}
}
\keyword{internal}
