% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translator.R
\name{sdm_to_bo}
\alias{sdm_to_bo}
\title{Translate file name from sdmpredictors to bio-oracle}
\usage{
sdm_to_bo(sdm)
}
\arguments{
\item{sdm}{the name of the sdm string}
}
\value{
A string with the name of the same file in bio-oracle
}
\description{
Translate file name from sdmpredictors to bio-oracle
}
\examples{
sdm_to_bo("BO2_tempmean_ss.tif")
sdm_to_bo("BO21_RCP26_2050_curvelltmax_bdmax.tif")
}
