% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\docType{data}
\name{equalareaproj}
\alias{equalareaproj}
\title{World Behrmann equal area coordinate reference system (ESRI:54017), used when
using load_layers with equal_area = TRUE}
\format{An object of class \code{CRS} of length 1.}
\usage{
equalareaproj
}
\description{
World Behrmann equal area coordinate reference system (ESRI:54017), used when
using load_layers with equal_area = TRUE
}
\keyword{datasets}

