% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get_datadir}
\alias{get_datadir}
\title{Get data directory}
\usage{
get_datadir(datadir)
}
\arguments{
\item{datadir}{character. Directory as passed to \code{\link{load_layers}}.
This does not change the data directory used by \code{\link{load_layers}}
but only shows the result of passing a specific datadir. If \code{NULL} is
passed then the \code{sdmpredictors_datadir} option is read. To set this
run \code{options(sdmpredictors_datadir = "<your preferred directory>")} in
every session or in your .RProfile.}
}
\value{
Path to the data directory.
}
\description{
Find out where the environmental data is stored
}
\seealso{
\code{\link{load_layers}}
}
\keyword{internal}
