% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\encoding{UTF-8}
\name{layers_correlation}
\alias{layers_correlation}
\title{Gives the Pearson correlation between layers}
\usage{
layers_correlation(layercodes = c(), include_quadratic = TRUE)
}
\arguments{
\item{layercodes}{character vector or dataframe. Codes of the layers you want
the basic statistics of as a character vector or a dataframe with a "layer_code" column.
With the default empty vector all statistics are returned.}

\item{include_quadratic}{logical. When \code{TRUE}, then the correlation coefficients of
the square of the layercodes area also returned. These layers are indicated with the
layercode with an extra ² symbol, for example \code{"BO_calcite"} became \code{"BO_calcite²"}.}
}
\value{
A dataframe with the Pearson product-moment correlation coefficients.
}
\description{
\code{layers_correlations} returns the Pearson product-moment correlation coefficient
(Pearson's r) for every combination of the give layercodes. The correlation between
a terrestrial and a marine layer has been set to \code{NA}.
}
\examples{
layers_correlation()
layers_correlation(c("BO_calcite", "MS_bathy_5m"))
layers_correlation(c("BO_calcite", "MS_bathy_5m"), include_quadratic = FALSE)
}
\seealso{
\code{\link{list_layers}} \code{\link{layer_stats}} \code{\link{correlation_groups}}
}

