% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\docType{data}
\name{lonlatproj}
\alias{lonlatproj}
\title{Longitude/latitude coordinate reference system (EPSG:4326), used when using
load_layers with equal_area = FALSE}
\format{
An object of class \code{CRS} of length 1.
}
\usage{
lonlatproj
}
\description{
Longitude/latitude coordinate reference system (EPSG:4326), used when using
load_layers with equal_area = FALSE
}
\keyword{datasets}
